把生成机器码的功能重构到一个新的包中
