+++ APDPlat_Core/src/main/java/org/apdplat/module/security/service/sequence/SequenceService.java
@@ -18,21 +18,18 @@
  * 
  */
 
-package org.apdplat.module.security.service;
+package org.apdplat.module.security.service.sequence;
 
 /**
- *在Mac OS X平台上生成机器码
+ *生成机器码的接口，不同平台有不同实现
  * @author 杨尚川
  */
-public class MacSequenceService    extends AbstractSequenceService{
-    @Override
-    public String getSequence() {
-        return getSigarSequence("mac");
-    }
-   
-    public static void main(String[] args) {
-        MacSequenceService s = new MacSequenceService();
-        String seq = s.getSequence();
-        System.out.println(seq);
-    }
+public interface SequenceService {
+
+    /**
+     * 获取机器码
+     * @return  机器码
+     */
+    public String getSequence();
+    
 }
