+++ APDPlat_Core/src/main/java/org/apdplat/module/security/service/sequence/SolarisSequenceService.java
@@ -18,21 +18,24 @@
  * 
  */
 
-package org.apdplat.module.security.service;
+package org.apdplat.module.security.service.sequence;
+
+import org.apdplat.module.security.service.sequence.AbstractSequenceService;
 
 /**
- *在Linux平台上生成机器码
+ *在Solaris平台上生成机器码
  * @author 杨尚川
  */
-public class LinuxSequenceService  extends AbstractSequenceService{
+public class SolarisSequenceService    extends AbstractSequenceService{
     @Override
     public String getSequence() {
-        return getSigarSequence("linux");
+        return getSigarSequence("solaris");
     }
 
     public static void main(String[] args) {
-        LinuxSequenceService s = new LinuxSequenceService();
+        SolarisSequenceService s = new SolarisSequenceService();
         String seq = s.getSequence();
         System.out.println(seq);
     }
+    
 }
